+++ R/fit_helpers.R
@@ -225,8 +225,10 @@ make_cens_prob_model <- function(obj, eval_env) {
   if (!is.null(eval_env$weights)) {
     cl <- rlang::call_modify(cl, caseweights = rlang::expr(eval_env$weights))
   }
-  rkm <- rlang::eval_tidy(cl)
-  attr(formula, ".Environment") <- rlang::base_env()
+  rkm <- try(rlang::eval_tidy(cl), silent = TRUE)
+  if (!inher(rkm, "try-error")) {
+    attr(formula, ".Environment") <- rlang::base_env()
+  }
   attr(rkm$formula, ".Environment") <- rlang::base_env()
   list(formula = formula, model = rkm)
 }